Revert "feat: add logic screen detail badge"
This reverts commit d22014181fe53740c5d238ae2e73aeab3c41ad59.
