+++ src/app/app.module.ts
@@ -76,11 +76,11 @@ import {
 } from './components/UnidadeCurricular/encontros/encontros.component';
 
 import {
-  ListagemTabelaBadgesComponent,
+  ListagemBadgesComponent,
   DialogExclusaoBadgeComponent
-} from './components/Dashboard/usuario-dashboard/usuario-badges/professor/listagem-tabela-badges/listagem-tabela-badges.component';
-import {NovoBadgeComponent} from './components/Dashboard/usuario-dashboard/usuario-badges/professor/novo-badge/novo-badge.component';
-import {AtualizarBadgeComponent} from './components/Dashboard/usuario-dashboard/usuario-badges/professor/atualizar-badge/atualizar-badge.component';
+} from './components/Badge/listagem-badge/listagem-badges.component';
+import {NovoBadgeComponent} from './components/Badge/novo-badge/novo-badge.component';
+import {AtualizarBadgeComponent} from './components/Badge/atualizar-badge/atualizar-badge.component';
 
 import {MenuUsuarioComponent} from './components/Dashboard/header/menu-usuario/menu-usuario.component';
 
@@ -100,11 +100,11 @@ import {UsuarioDadosComponent} from './components/Dashboard/usuario-dashboard/us
 import {
   UsuarioProximasAtividadesComponent
 } from './components/Dashboard/usuario-dashboard/usuario-proximas-atividades/usuario-proximas-atividades.component';
-import {ListagemUltimasBadgesComponent} from './components/Dashboard/usuario-dashboard/usuario-badges/estudante/listagem-ultimas-badges/listagem-ultimas-badges.component';
+import {UsuarioBadgesComponent} from './components/Dashboard/usuario-dashboard/usuario-badges/usuario-badges.component';
 
 import {
-  ListagemBadgesComponent
-} from './components/Dashboard/usuario-dashboard/usuario-badges/estudante/listagem-badges/listagem-badges.component';
+  ListagemUsuarioBadgesComponent
+} from './components/Dashboard/usuario-dashboard/usuario-badges/listagem-usuario-badges/listagem-usuario-badges.component';
 import {ApoioDuvidasComponent} from './components/ApoioDuvidas/apoio-duvidas.component';
 import {ComentarioComponent} from './components/ApoioDuvidas/comentario/comentario.component';
 
@@ -225,8 +225,6 @@ import {
 import {TagModule} from "primeng/tag";
 import {ProgressSpinnerModule} from "primeng/progressspinner";
 import {MessageService} from "primeng/api";
-import { UsuarioBadgesComponent } from './components/Dashboard/usuario-dashboard/usuario-badges/usuario-badges.component';
-import { DetalheBadgeComponent } from './components/Dashboard/usuario-dashboard/usuario-badges/estudante/detalhe-badge/detalhe-badge.component';
 
 registerLocaleData(localePT);
 registerLocaleData(localeES);
@@ -276,8 +274,8 @@ registerLocaleData(localeFR);
     UsuarioArquivosPrivadosComponent,
     UsuarioDadosComponent,
     UsuarioProximasAtividadesComponent,
-    ListagemUltimasBadgesComponent,
-    ListagemBadgesComponent,
+    UsuarioBadgesComponent,
+    ListagemUsuarioBadgesComponent,
     ApoioDuvidasComponent,
     ComentarioComponent,
     ListaObjetosAprendizagem,
@@ -293,10 +291,7 @@ registerLocaleData(localeFR);
     UsuarioNoticiasComponent,
     CardMensagemComponent,
     SalvarRegistroAvaliacaoComponent,
-    ListarRegistrosAvaliacoesComponent,
-    ListagemTabelaBadgesComponent,
-    UsuarioBadgesComponent,
-    DetalheBadgeComponent
+    ListarRegistrosAvaliacoesComponent
   ],
   imports: [
     BrowserModule,
