+++ src/app/components/Badge/listagem-badge/listagem-badges.component.html
@@ -1,0 +1,81 @@
+<div *ngIf="!badges">
+    <mat-progress-bar mode="buffer"></mat-progress-bar>
+</div>
+<div class="dadosBadges">
+    <div class="cabecalho">
+        <mat-grid-list cols="2" rowHeight="50px">
+            <mat-grid-tile>
+                <h2 class="titulo">Badges Cadastrados</h2>
+            </mat-grid-tile>            
+            <mat-grid-tile>
+                <button mat-stroked-button color="primary" routerLink="/badges/novobadge">Novo Badge</button>
+            </mat-grid-tile>
+        </mat-grid-list>
+        <mat-divider></mat-divider>
+
+        <div class="divAutoComplete">
+            <form>
+                <mat-form-field class="autoCompleteInput">
+                    <input 
+                        type="text" 
+                        placeholder="Buscar" 
+                        matInput 
+                        [formControl]="autoCompleteInput"
+                        [matAutocomplete]="auto"
+                    />
+                    <mat-autocomplete autoActiveFirstOption #auto="matAutocomplete">
+                        <mat-option *ngFor="let descricao of descricaoBadges | async" [value]="descricao">
+                            {{descricao}}
+                        </mat-option>
+                    </mat-autocomplete>
+                </mat-form-field>
+            </form>
+        </div>
+
+        <table *ngIf="badges"
+            mat-table 
+            [dataSource]="badges"
+            matSort
+            class="mat-elevation-z8 tabela"
+        >
+            <ng-container matColumnDef="id">
+                <th mat-header-cell *matHeaderCellDef>Id</th>
+                <td mat-cell *matCellDef="let badge">{{ badge.id }}</td>
+            </ng-container>
+            <ng-container matColumnDef="descricao">
+                <th mat-header-cell *matHeaderCellDef mat-sort-header>Descrição</th>
+                <td mat-cell *matCellDef="let badge">{{ badge.descricao }}</td>
+            </ng-container>
+            
+            <ng-container matColumnDef="imagem">
+                <th mat-header-cell *matHeaderCellDef mat-sort-header>Imagem</th>
+                <td mat-cell *matCellDef="let badge"></td>
+            </ng-container>
+
+            <ng-container matColumnDef="badgeNivel">
+                <th mat-header-cell *matHeaderCellDef mat-sort-header>Nivel</th>
+                <td mat-cell *matCellDef="let badge">{{ badge.badgeNivel.descricao }}</td>
+            </ng-container>
+
+            <ng-container matColumnDef="acoes">
+                <th mat-header-cell *matHeaderCellDef>Ações</th>
+                <td mat-cell *matCellDef="let badge">
+                    <button mat-mini-fab color="primary" [routerLink]="['/badges/atualizarbadge', badge.id]">
+                        <mat-icon>edit</mat-icon>
+                    </button>
+                    <button mat-mini-fab color="warn" (click)="AbrirDialog(badge.id, badge.descricao)">
+                        <mat-icon>delete_forever</mat-icon>
+                    </button>
+                </td>
+            </ng-container>
+
+            <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+            <tr mat-row *matRowDef="let row columns: displayedColumns"></tr>
+
+        </table>
+        <mat-paginator
+            [pageSizeOptions]="[5, 10, 15, 20]"
+            showFirstLastButtons
+        ></mat-paginator>
+    </div>
+</div>
