+++ src/app/components/Dashboard/dashboard/mensagens/mensagens.component.html
@@ -47,7 +47,7 @@
         <div class="card-body" >
           <ul class="list-unstyled">
             <li *ngFor="let item of mensagens" >
-              <app-card-mensagem
+              <app-card-mensagem 
               [mensagem]="item"
               ></app-card-mensagem>
             </li>
