--- src/app/components/Dashboard/usuario-dashboard/usuario-badges/estudante/detalhe-badge/detalhe-badge.component.html
@@ -1,1 +1,0 @@
-<p>detalhe-badge works!</p>
