--- src/app/components/Dashboard/usuario-dashboard/usuario-badges/estudante/listagem-badges/listagem-badges.component.css
@@ -1,10 +1,0 @@
-.card {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.image {
-  max-width: 120px;
-}
