+++ src/app/components/Dashboard/usuario-dashboard/usuario-badges/listagem-usuario-badges/listagem-usuario-badges.component.css
@@ -1,0 +1,102 @@
+/* CARD JAVA */
+.title-java {
+    margin-left: 80px;
+}
+
+.img-java {
+    max-width: 120px;
+    left: 25%;
+}
+
+.card-java {
+    position: absolute;
+    left: 20%;
+    top: 15%;
+    max-width: 250px;
+
+}
+
+/* CARD PYTON */
+.title-pyton {
+    margin-left: 80px;
+}
+
+.img-pyton {
+    max-width: 110px;
+    left: 25%;
+}
+
+.card-pyton {
+    position: absolute;
+    left: 45%;
+    top: 15%;
+    max-width: 250px;
+}
+
+/* CARD HTML */
+.title-html {
+    margin-left: 80px;
+}
+
+.img-html {
+    max-width: 120px;
+    left: 25%;
+}
+
+.card-html {
+    position: absolute;
+    left: 70%;
+    top: 15%;
+    max-width: 250px;
+}
+
+/* CARD CSS */
+.title-css {
+    margin-left: 80px;
+}
+
+.img-css {
+    max-width: 85px;
+    left: 25%;
+}
+
+.card-css {
+    position: absolute;
+    left: 20%;
+    top: 70%;
+    max-width: 250px;
+}
+
+/* CARD JAVASCRIPT */
+.title-js {
+    margin-left: 60px;
+}
+
+.img-js {
+    max-width: 120px;
+    left: 25%;
+}
+
+.card-js {
+    position: absolute;
+    left: 45%;
+    top: 70%;
+    max-width: 250px;
+}
+
+/* CARD ANGULAR */
+.title-angular {
+    margin-left: 60px;
+}
+
+.img-angular {
+    max-width: 120px;
+    left: 25%;
+}
+
+.card-angular {
+    position: absolute;
+    left: 70%;
+    top: 70%;
+    max-width: 250px;
+}
