--- src/app/components/Dashboard/usuario-dashboard/usuario-badges/professor/atualizar-badge/atualizar-badge.component.html
@@ -1,103 +1,0 @@
-<div *ngIf="!formulario">
-    <mat-progress-bar mode="buffer"></mat-progress-bar>
-</div>
-<div class="divCard" *ngIf="formulario">
-    <mat-card>
-        <mat-card-title>
-            <h3>Atualizar Badge : {{descricaoBadge}}</h3>
-        </mat-card-title>
-
-        <form [formGroup]="formulario" (ngSubmit)="EnviarFormulario()">
-            <div class="cardErros" *ngIf="erros.length > 0">
-                <mat-card>
-                    <mat-card-header>
-                        <mat-card-title>Erro no formulário!</mat-card-title>
-                    </mat-card-header>
-                    <mat-divider></mat-divider>
-                    <mat-card-content>
-                        <div class="divErros">
-                            <h4 class="textoErros" *ngFor="let erro of erros">
-                                {{erro}}
-                            </h4>
-                        </div>
-                    </mat-card-content>
-                </mat-card>
-            </div>            
-            <mat-card-content>
-                <div>
-                    <mat-form-field>
-                        <mat-label>
-                            Descrição
-                        </mat-label>
-                        <input 
-                            type="text" 
-                            matInput 
-                            formControlName="descricao" 
-                            placeholder="Digite a descrição do Badge"
-                            required
-                        />
-                        <mat-error *ngIf="propriedade.descricao.touched && propriedade.descricao.invalid">
-                            <sup>*</sup>Digite a descrição do Badge
-                        </mat-error>
-                    </mat-form-field>
-                </div>
-                <div>
-                    <div class="divFoto">
-                        <img hidden [src]="imagem" src="#" width="200" height="200" />                                
-                    </div>
-                    <input 
-                        hidden 
-                        type="file" 
-                        #fileInput 
-                        name="imagem" 
-                        formControlName="imagem" 
-                        required 
-                        (change)="SelecionarImagem($event)"
-                    />
-                    <br/><br/><br/><br/>
-                    <button type="button" mat-stroked-button color="primary" (click)="fileInput.click()">
-                        Escolha a imagem do Badge
-                    </button>
-
-                    <div *ngIf="propriedade.imagem.touched && propriedade.imagem.invalid">
-                        <sup>*</sup>Escolha a imagem do Badge.
-                    </div>
-                </div>
-                <div>
-                    <mat-form-field>
-                        <mat-label>
-                            Selecione o nivel do Badge
-                        </mat-label>
-                        <mat-select formControlName="badgeNivelId" required="">
-                            <mat-option *ngFor="let nivel of badgeNivel" [value]="nivel.id">{{nivel.descricao}}</mat-option>
-                        </mat-select>
-                        <mat-error *ngIf="propriedade.badgeNivelId.touched && propriedade.badgeNivelId.invalid">
-                            <sup>*</sup>Selecione o nivel do Badge
-                        </mat-error>                        
-                    </mat-form-field>
-                </div>
-                <div>
-                    <mat-form-field>
-                        <mat-label>
-                            Status de Operação
-                        </mat-label>
-                        <mat-select formControlName="status" required>
-                            <mat-option *ngFor="let estado of estados" [value]="estado.value">
-                                {{estado.viewValue}}
-                            </mat-option>
-                        </mat-select>
-                    </mat-form-field>
-                </div>
-            </mat-card-content>
-            <mat-divider></mat-divider>
-            <mat-card-actions>
-                <div>
-                    <button type="submit" mat-stroked-button color="primary" [disabled]="formulario.invalid">
-                        Atualizar
-                    </button>
-                    <button mat-stroked-button type="button" (click)="VoltarListagem()">Voltar</button>
-                </div>
-            </mat-card-actions>
-        </form>
-    </mat-card>
-</div>
