--- src/app/components/Dashboard/usuario-dashboard/usuario-badges/professor/listagem-tabela-badges/dialog-exclusao-badge.html
@@ -1,6 +1,0 @@
-<h1 mat-dialog-title>Confirmação de Exclusão</h1>
-<div mat-dialog-content>Deseja prosseguir com a exclusão do badge {{ dados.nome}} ?</div>
-<mat-dialog-actions align="end">
-    <button mat-button mat-dialog-close>Não</button>
-    <button mat-button [mat-dialog-close]="true" (click)="ExcluirBadge(dados.id)">Sim</button>
-</mat-dialog-actions>
