--- src/app/components/Dashboard/usuario-dashboard/usuario-badges/professor/novo-badge/novo-badge.component.html
@@ -1,119 +1,0 @@
-<div *ngIf="!formulario">
-    <mat-progress-bar mode="buffer"></mat-progress-bar>
-</div>
-<div class="divCard" *ngIf="formulario">
-    <mat-card>
-        <mat-card-title>
-            <h3>Cadastro de Badges</h3>
-        </mat-card-title>
-
-        <form [formGroup]="formulario" (ngSubmit)="EnviarFormulario()">
-            <div class="cardErros" *ngIf="erros.length > 0">
-                <mat-card>
-                    <mat-card-header>
-                        <mat-card-title>Erro no formulário!</mat-card-title>
-                    </mat-card-header>
-                    <mat-divider></mat-divider>
-                    <mat-card-content>
-                        <div class="divErros">
-                            <h4 class="textoErros" *ngFor="let erro of erros">
-                                {{erro}}
-                            </h4>
-                        </div>
-                    </mat-card-content>
-                </mat-card>
-            </div>
-            <mat-card-content>
-                <div>
-                    <mat-form-field>
-                        <mat-label>
-                            Descrição
-                        </mat-label>
-                        <input 
-                            type="text" 
-                            matInput 
-                            formControlName="descricao" 
-                            placeholder="Digite a descrição do Badge"
-                            required
-                        />
-                        <mat-error *ngIf="propriedade.descricao.touched && propriedade.descricao.invalid">
-                            <sup>*</sup>Digite a descrição do Badge
-                        </mat-error>
-                    </mat-form-field>
-                </div>
-                <div>
-                    <div class="divFoto">
-                        <img hidden id="imagem" src="#" width="200" height="200" />                                
-                    </div>
-                    <input 
-                        hidden 
-                        type="file" 
-                        #fileInput 
-                        name="imagem" 
-                        formControlName="imagem" 
-                        required 
-                        (change)="SelecionarImagem($event)"
-                    />
-                    <br/><br/><br/><br/>
-                    <button type="button" mat-stroked-button color="primary" (click)="fileInput.click()">
-                        Escolha a imagem do Badge
-                    </button>
-
-                    <div *ngIf="propriedade.imagem.touched && propriedade.imagem.invalid">
-                        <sup>*</sup>Escolha a imagem do Badge.
-                    </div>
-
-                    <!--
-                    <mat-form-field>
-                        <mat-label>
-                            Imagem
-                        </mat-label>
-                        <input 
-                            type="text" 
-                            matInput 
-                            formControlName="imagem" 
-                            placeholder="Selecione a imagem do badge"
-                            required
-                        />
-                        <mat-error *ngIf="propriedade.imagem.touched && propriedade.imagem.invalid">
-                            <sup>*</sup>Selecione a imagem do badge
-                        </mat-error>
-                    </mat-form-field> -->
-                </div>
-                <div>
-                    <mat-form-field>
-                        <mat-label>
-                            Selecione o Nivel do Badge
-                        </mat-label>
-                        <mat-select formControlName="badgeNivelId" required="">
-                            <mat-option *ngFor="let nivel of badgesNivel" [value]="nivel.id">{{nivel.descricao}}</mat-option>
-                        </mat-select>
-                        <mat-error *ngIf="propriedade.badgeNivelId.touched && propriedade.badgeNivelId.invalid">
-                            <sup>*</sup>Selecione o Nivel do Badge
-                        </mat-error>
-                    </mat-form-field>
-                </div>
-                <div>
-                    <mat-form-field>
-                        <mat-label>
-                            Status de Operação
-                        </mat-label>
-                        <mat-select formControlName="status">
-                            <mat-option value='1'>Ativo</mat-option>
-                            <mat-option value='0'>Inativo</mat-option>
-                        </mat-select>
-                    </mat-form-field>
-                </div>                
-            </mat-card-content>
-            <mat-divider></mat-divider>
-            <mat-card-actions>
-                <div>
-                    <button type="submit" mat-stroked-button color="primary" [disabled]="formulario.invalid">
-                        Salvar
-                    </button>
-                    <button mat-stroked-button type="button" (click)="VoltarListagem()">Voltar</button>
-                </div>
-            </mat-card-actions>
-        </form>
-    </mat-card>
-</div>
