+++ src/app/components/Dashboard/usuario-dashboard/usuario-badges/usuario-badges.component.css
@@ -1,0 +1,43 @@
+*,
+*::before,
+*::after {
+	padding: 0;
+	margin: 0;
+	box-sizing: border-box;
+}
+
+.card {
+	width: 20rem;
+	border-radius: 1rem;
+	background: white;
+	box-shadow: 8px 3px 22px 10px rgba(150, 150, 150, 0.11);
+	border: transparent;
+	position: relative;
+	color: #434343;
+}
+
+.card__container {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	padding: 0rem;
+	width: 100%;
+	height: 100%;
+	background: white;
+	border-radius: 1rem;
+}
+
+.card .card__header {
+	margin-bottom: 3rem;
+	color: #092c4c;
+}
+
+.card__body {
+	padding: 0px 10px;
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	width: 80%;
+	scale: 1.5;
+	color: orange;
+}
