+++ src/app/components/Dashboard/usuario-dashboard/usuario-badges/usuario-badges.component.html
@@ -1,1 +1,18 @@
-<p>usuario-badges works!</p>
+<div class="">
+	<div class="card__container" >
+		<h2 class="card__header">Últimos 3 badges</h2>
+		<div class="card__body">
+			<img  style="width: 3rem" src="{{'data:image/png;base64,' + badges[2].imagem}}"/>
+			<img  style="width: 3rem" src="{{'data:image/png;base64,' + badges[0].imagem}}"/>
+			<img  style="width: 3rem" src="{{'data:image/png;base64,' + badges[1].imagem}}"/>
+		</div>
+		<br>
+		<button
+      mat-mini-fab
+      color="primary"
+      [routerLink]="['/badges/listagemusuariobadges']"
+    >
+		<mat-icon>list</mat-icon>
+		</button>
+	</div>
+</div>
