+++ src/app/components/Dashboard/usuario-dashboard/usuario-badges/usuario-badges.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Badge } from 'src/app/models/Badge';
+import { AuthGuardService } from 'src/app/services/auth-guard.service';
+import { BadgeService } from 'src/app/services/badge.service';
 
 @Component({
   selector: 'app-usuario-badges',
@@ -6,10 +9,29 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./usuario-badges.component.css']
 })
 export class UsuarioBadgesComponent implements OnInit {
+  badges: Badge[]=[];
+  badge : Badge = new Badge;
+  idUsuarioLogado: string;
 
-  constructor() { }
+
+  constructor(
+    private badgeService: BadgeService,
+    private authGuardService: AuthGuardService
+  ) { }
 
   ngOnInit(): void {
+
+    this.idUsuarioLogado = this.authGuardService.getIdUsuarioLogado();
+
+    this.badgeService.ObterBadgesRecentesPeloUsuarioId(this.idUsuarioLogado).subscribe(resultado => {
+      this.badges = resultado;
+    })
+  }
+
+  getImage(baseImage: string): any {
+
+    let objectURL = 'data:image/png;base64,' + atob(baseImage);
+    return objectURL;
   }
 
 }
