+++ src/app/components/Dashboard/usuario-dashboard/usuario-dashboard.component.html
@@ -4,7 +4,7 @@
       <div class="dashboard-usuario-card" style="display: flex; flex-direction: column; justify-content: space-around;">
         <app-usuario-dados></app-usuario-dados>
         <p-divider type="solid"></p-divider>
-        <app-listagem-ultimas-badges></app-listagem-ultimas-badges>
+        <app-usuario-badges></app-usuario-badges>
       </div>
 	  </div>
 	  <div>
